(refractor) Change of folder name
