+++ server/services/reports/src/main/java/com/recykling/report/urtReport/dto/UrtReportDto.java
@@ -1,18 +1,18 @@
 package com.recykling.report.urtReport.dto;
 
-import com.recykling.report.urtReport.entities.urtAlCuRefrigerator.AlCuRefrigeratorWeights;
-import com.recykling.report.urtReport.entities.urtAlCuRefrigerator.dto.AlCuRefrigeratorWeightsDto;
-import com.recykling.report.urtReport.entities.urtOilFromAggregatesWeights.OilFromAggregatesWeights;
-import com.recykling.report.urtReport.entities.urtOilFromAggregatesWeights.dto.OilFromAggregatesWeightsDto;
-import com.recykling.report.urtReport.entities.urtPsAbsRefrigeratorWeights.PsAbsRefrigeratorWeights;
-import com.recykling.report.urtReport.entities.urtPsAbsRefrigeratorWeights.dto.PsAbsRefrigeratorWeightsDto;
-import com.recykling.report.urtReport.entities.urtRefrigeratorPowerCableWeights.RefrigeratorPowerCableWeights;
-import com.recykling.report.urtReport.entities.urtRefrigeratorPowerCableWeights.dto.RefrigeratorPowerCableWeightsDto;
-import com.recykling.report.urtReport.entities.urtAggregatesWithoutOilWeights.dto.AggregatesWithoutOilDto;
+import com.recykling.report.urtReport.manyToOne.urtAlCuRefrigerator.AlCuRefrigeratorWeights;
+import com.recykling.report.urtReport.manyToOne.urtAlCuRefrigerator.dto.AlCuRefrigeratorWeightsDto;
+import com.recykling.report.urtReport.manyToOne.urtOilFromAggregatesWeights.OilFromAggregatesWeights;
+import com.recykling.report.urtReport.manyToOne.urtOilFromAggregatesWeights.dto.OilFromAggregatesWeightsDto;
+import com.recykling.report.urtReport.manyToOne.urtPsAbsRefrigeratorWeights.PsAbsRefrigeratorWeights;
+import com.recykling.report.urtReport.manyToOne.urtPsAbsRefrigeratorWeights.dto.PsAbsRefrigeratorWeightsDto;
+import com.recykling.report.urtReport.manyToOne.urtRefrigeratorPowerCableWeights.RefrigeratorPowerCableWeights;
+import com.recykling.report.urtReport.manyToOne.urtRefrigeratorPowerCableWeights.dto.RefrigeratorPowerCableWeightsDto;
+import com.recykling.report.urtReport.manyToOne.urtAggregatesWithoutOilWeights.dto.AggregatesWithoutOilDto;
 import com.recykling.report.employee.dto.EmployeeDto;
-import com.recykling.report.urtReport.entities.urtReportHistory.UrtReportHistory;
+import com.recykling.report.urtReport.manyToOne.urtReportHistory.UrtReportHistory;
 import com.recykling.report.employee.Employee;
-import com.recykling.report.urtReport.entities.urtAggregatesWithoutOilWeights.AggregatesWithoutOilWeights;
+import com.recykling.report.urtReport.manyToOne.urtAggregatesWithoutOilWeights.AggregatesWithoutOilWeights;
 import com.recykling.report.valueObjects.*;
 import lombok.Data;
 import lombok.NoArgsConstructor;
