--- server/services/reports/src/main/java/com/recykling/report/urtReport/entities/urtAggregatesWithoutOilWeights/repo/AggregatesWithoutOilWeightsRepository.java
@@ -1,9 +1,0 @@
-package com.recykling.report.urtReport.entities.urtAggregatesWithoutOilWeights.repo;
-
-import com.recykling.report.urtReport.entities.urtAggregatesWithoutOilWeights.AggregatesWithoutOilWeights;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface AggregatesWithoutOilWeightsRepository extends JpaRepository<AggregatesWithoutOilWeights, Long> {
-}
