--- server/services/reports/src/main/java/com/recykling/report/urtReport/entities/urtAluminium/repo/AluminiumWeightsRepository.java
@@ -1,9 +1,0 @@
-package com.recykling.report.urtReport.entities.urtAluminium.repo;
-
-import com.recykling.report.urtReport.entities.urtAluminium.AluminiumWeights;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface AluminiumWeightsRepository extends JpaRepository<AluminiumWeights, Long> {
-}
