--- server/services/reports/src/main/java/com/recykling/report/urtReport/entities/urtOilFromAggregatesWeights/repo/OilFromAggregatesWeightsRepository.java
@@ -1,9 +1,0 @@
-package com.recykling.report.urtReport.entities.urtOilFromAggregatesWeights.repo;
-
-import com.recykling.report.urtReport.entities.urtOilFromAggregatesWeights.OilFromAggregatesWeights;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface OilFromAggregatesWeightsRepository extends JpaRepository<OilFromAggregatesWeights, Long> {
-}
