--- server/services/reports/src/main/java/com/recykling/report/urtReport/entities/urtRefrigeratorPowerCableWeights/dto/RefrigeratorPowerCableWeightsDto.java
@@ -1,21 +1,0 @@
-package com.recykling.report.urtReport.entities.urtRefrigeratorPowerCableWeights.dto;
-
-import com.recykling.report.urtReport.entities.urtRefrigeratorPowerCableWeights.RefrigeratorPowerCableWeights;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.ArrayList;
-import java.util.List;
-
-@Data @NoArgsConstructor
-public class RefrigeratorPowerCableWeightsDto {
-    private List<Integer> refrigeratorPowerCableWeights = new ArrayList<>();
-    private Integer count;
-
-    public RefrigeratorPowerCableWeightsDto(List<RefrigeratorPowerCableWeights> weights) {
-        weights
-               .forEach(weight ->
-                       refrigeratorPowerCableWeights.add(weight.getWeight()));
-       this.count = refrigeratorPowerCableWeights.size();
-    }
-}
