+++ server/services/reports/src/main/java/com/recykling/report/urtReport/manyToOne/urtOilFromAggregatesWeights/dto/OilFromAggregatesWeightsDto.java
@@ -1,0 +1,21 @@
+package com.recykling.report.urtReport.manyToOne.urtOilFromAggregatesWeights.dto;
+
+import com.recykling.report.urtReport.manyToOne.urtOilFromAggregatesWeights.OilFromAggregatesWeights;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@Data @NoArgsConstructor
+public class OilFromAggregatesWeightsDto {
+    private List<Integer> oilFromAggregatesWeights = new ArrayList<>();
+    private Integer count;
+
+    public OilFromAggregatesWeightsDto(List<OilFromAggregatesWeights> weights) {
+        weights
+               .forEach(weight ->
+                       oilFromAggregatesWeights.add(weight.getWeight()));
+       this.count = oilFromAggregatesWeights.size();
+    }
+}
