+++ server/services/reports/src/main/java/com/recykling/report/urtReport/service/impl/UrtReportServiceImpl.java
@@ -1,18 +1,18 @@
 package com.recykling.report.urtReport.service.impl;
 
-import com.recykling.report.urtReport.entities.urtAlCuRefrigerator.AlCuRefrigeratorWeights;
-import com.recykling.report.urtReport.entities.urtAlCuRefrigerator.repo.AlCuRefrigeratorWeightsRepository;
-import com.recykling.report.urtReport.entities.urtOilFromAggregatesWeights.OilFromAggregatesWeights;
-import com.recykling.report.urtReport.entities.urtOilFromAggregatesWeights.repo.OilFromAggregatesWeightsRepository;
-import com.recykling.report.urtReport.entities.urtPsAbsRefrigeratorWeights.PsAbsRefrigeratorWeights;
-import com.recykling.report.urtReport.entities.urtPsAbsRefrigeratorWeights.repo.PsAbsRefrigeratorWeightsRepository;
-import com.recykling.report.urtReport.entities.urtRefrigeratorPowerCableWeights.RefrigeratorPowerCableWeights;
-import com.recykling.report.urtReport.entities.urtRefrigeratorPowerCableWeights.repo.RefrigeratorPowerCableWeightsRepository;
+import com.recykling.report.urtReport.manyToOne.urtAlCuRefrigerator.AlCuRefrigeratorWeights;
+import com.recykling.report.urtReport.manyToOne.urtAlCuRefrigerator.repo.AlCuRefrigeratorWeightsRepository;
+import com.recykling.report.urtReport.manyToOne.urtOilFromAggregatesWeights.OilFromAggregatesWeights;
+import com.recykling.report.urtReport.manyToOne.urtOilFromAggregatesWeights.repo.OilFromAggregatesWeightsRepository;
+import com.recykling.report.urtReport.manyToOne.urtPsAbsRefrigeratorWeights.PsAbsRefrigeratorWeights;
+import com.recykling.report.urtReport.manyToOne.urtPsAbsRefrigeratorWeights.repo.PsAbsRefrigeratorWeightsRepository;
+import com.recykling.report.urtReport.manyToOne.urtRefrigeratorPowerCableWeights.RefrigeratorPowerCableWeights;
+import com.recykling.report.urtReport.manyToOne.urtRefrigeratorPowerCableWeights.repo.RefrigeratorPowerCableWeightsRepository;
 import com.recykling.report.urtReport.service.IUrtReportService;
-import com.recykling.report.urtReport.entities.urtReportHistory.UrtReportHistory;
-import com.recykling.report.urtReport.entities.urtAggregatesWithoutOilWeights.AggregatesWithoutOilWeights;
-import com.recykling.report.urtReport.entities.urtAggregatesWithoutOilWeights.repo.AggregatesWithoutOilWeightsRepository;
-import com.recykling.report.urtReport.entities.urtReportHistory.repo.UrtReportHistoryRepository;
+import com.recykling.report.urtReport.manyToOne.urtReportHistory.UrtReportHistory;
+import com.recykling.report.urtReport.manyToOne.urtAggregatesWithoutOilWeights.AggregatesWithoutOilWeights;
+import com.recykling.report.urtReport.manyToOne.urtAggregatesWithoutOilWeights.repo.AggregatesWithoutOilWeightsRepository;
+import com.recykling.report.urtReport.manyToOne.urtReportHistory.repo.UrtReportHistoryRepository;
 import com.recykling.report.urtReport.controller.request.RequestCreateUrtReport;
 import com.recykling.report.urtReport.dto.UrtReportDto;
 import com.recykling.report.urtReport.UrtReport;
